# Micro-optimize finding executables